feat(examples): demokit v0.0.16 + two-process server/client split
Bumps demokit v0.0.1 → v0.0.16 and migrates all 10 examples to the new
StepContext-based Run signature. Each example splits into:

- main.go         — boots the auth/resource servers. With --serve they
                    bind on real ports so any OAuth client (curl, your
                    app, an MCP host) can drive them.
- walkthrough.go  — the demokit demo, acting as one such client. Spins
                    up the same servers in-process via httptest using
                    the same builders --serve uses, so the wire bytes
                    are identical between modes.

Each example also ships a slim README.md (quick-start) plus a generated
WALKTHROUGH.md (mermaid sequence diagram + step-by-step + 35 copy-paste
curl reproductions across the 10 examples). WALKTHROUGH.md regenerates
via `make walkthrough` per example or `make walkthroughs` at the top
level.

New examples/common/ package centralizes:
- SetupRenderer(demo) — opt-in TUI renderer when --tui is passed
- NewOneAuthLogger(prefix, …) — canonical ColorLogger rules

Per-example Makefile targets: demo (default, TUI), demo-plain, demo-ci,
serve, walkthrough. Top-level adds `make walkthroughs` (regenerate all)
and `make demos-ci` (smoke-run every example non-interactively).

Also drops the old --readme flag and `make readmes` target — single
source of truth is the demo definition rendered via --doc md.

Updates CLAUDE.md and CAPABILITIES.md with the new conventions.
