+++ examples/02-resource-token-hs256/Makefile
@@ -1,16 +1,25 @@
 # 02 — Resource Token with HS256 (Federated Auth)
 #
-# Infrastructure: None (in-memory, self-contained)
+# main.go boots an auth server (registration only) and a resource
+# server. The walkthrough is a client that registers an app, mints
+# tokens locally, and uses them.
 
 EXAMPLE_DIR := 02-resource-token-hs256
 
-run:
+demo:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --tui
+
+demo-plain:
 	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/
 
-demo:
+demo-ci:
 	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --non-interactive
 
-readme:
-	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --readme > examples/$(EXAMPLE_DIR)/README.md
+serve:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --serve
 
-.PHONY: run demo readme
+walkthrough:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --doc md > examples/$(EXAMPLE_DIR)/WALKTHROUGH.md
+
+.PHONY: demo demo-plain demo-ci serve walkthrough
+.DEFAULT_GOAL := demo
