+++ examples/03-resource-token-rs256-jwks/Makefile
@@ -1,16 +1,21 @@
 # 03 — Resource Token with RS256 + JWKS Discovery
-#
-# Infrastructure: None (in-memory, self-contained)
 
 EXAMPLE_DIR := 03-resource-token-rs256-jwks
 
-run:
+demo:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --tui
+
+demo-plain:
 	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/
 
-demo:
+demo-ci:
 	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --non-interactive
 
-readme:
-	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --readme > examples/$(EXAMPLE_DIR)/README.md
+serve:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --serve
 
-.PHONY: run demo readme
+walkthrough:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --doc md > examples/$(EXAMPLE_DIR)/WALKTHROUGH.md
+
+.PHONY: demo demo-plain demo-ci serve walkthrough
+.DEFAULT_GOAL := demo
