+++ examples/06-dynamic-client-registration/Makefile
@@ -1,26 +1,21 @@
 # 06 — Dynamic Client Registration (RFC 7591)
-#
-# Infrastructure: None (Keycloak optional for cross-server DCR demo)
-#
-# To see the Keycloak step:
-#   make upkcl    # start Keycloak for examples (port 8280)
-#   make run      # the KC step auto-detects and runs
 
 EXAMPLE_DIR := 06-dynamic-client-registration
 
-run:
+demo:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --tui
+
+demo-plain:
 	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/
 
-demo:
+demo-ci:
 	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --non-interactive
 
-readme:
-	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --readme > examples/$(EXAMPLE_DIR)/README.md
+serve:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --serve
 
-upkcl:
-	$(MAKE) -C .. upkcl
+walkthrough:
+	cd ../.. && go run -buildvcs=false ./examples/$(EXAMPLE_DIR)/ --doc md > examples/$(EXAMPLE_DIR)/WALKTHROUGH.md
 
-downkcl:
-	$(MAKE) -C .. downkcl
-
-.PHONY: run demo readme upkcl downkcl
+.PHONY: demo demo-plain demo-ci serve walkthrough
+.DEFAULT_GOAL := demo
